- deleted ExceptionController duplicate - moved EventController to implementation package
