+++ src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/EventController.java
@@ -1,12 +1,12 @@
-package ro.societateahermes.backendservice.controller;
+package ro.societateahermes.backendservice.controller.controllerImplementation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ro.societateahermes.backendservice.service.EventServiceInterface;
 import ro.societateahermes.backendservice.entities.dto.NotificationSwitchDTO;
+import ro.societateahermes.backendservice.service.EventServiceInterface;
 
 @RestController
 @RequestMapping("/event")
